Major refactor
Eliminate the delination of "parent" and "sub" domain lists, all domains are sourced from the same list.
Added the concept of an "owner" list that is not explicitly scanned. It is meant to contain "break points" in the subdomain tree where an organization name is provided.
This was done in order to facilitate the use case where the owner of a parent domain (like canada.ca) was different than the owner of a subdomain of that parent (digital.canada.ca)